+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## master
+* Feature: (GUI) Added filtering by date to Logs v2
+* Feature: (GUI) Added hover and active states to the table in Logs v2
+* Feature: (GUI) Ability to reset sidebar width by double clicking the resize handle
+
 ## 4.15.2
 * Bug: Fixed running the GUI in custom port number