+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## master
+## 4.16.0
 * Feature: (GUI) Added filtering by date to Logs v2
 * Feature: (GUI) Added hover and active states to the table in Logs v2
 * Feature: (GUI) Ability to reset sidebar width by double clicking the resize handle